Refactor: inline the single-use server-variable regex and move the parsed-spec containment test
The pattern constant had one use eighteen lines below it and did not centralize
anything, since rest_api_tool.ts inlines the same literal. The containment case
asserts on prepareRequestParams, so it belongs with the other prepareRequestParams
cases rather than reaching across layers from the parser suite.
